Add installation instructions to the README
On copying the documentation from heroku I came accross installation instructions that seem to have gotten lost when the project came to github. I felt it would be useful to reintegrate them in the README. I don't know if there was a motivation for not bringing them in from heroku.